pyproject.toml
[project]
name = "Sipeed-NanoCluster-Server"
version = "0.1.0"
description = "This is a small server with a web interface for a Sipeed NanoCluster."
authors = [
    {name = "Stefan Dej", email = "meteyou@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "pigpio>=1.78",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "pigpio>=1.78",
]
